+++ packages/eas-cli/src/credentials/manager/Actions.ts
@@ -18,6 +18,7 @@ export enum AndroidActionType {
   GoBackToCaller,
   GoBackToHighLevelActions,
   CreateKeystore,
+  SetDefaultKeystore,
   DownloadKeystore,
   RemoveKeystore,
   CreateFcm,